+++ packages/core/src/ruleset/index.ts
@@ -1,7 +1,7 @@
-export { assertValidRuleset, RulesetValidationError } from './validation';
-export { getDiagnosticSeverity } from './utils';
-export { createRulesetFunction, SchemaDefinition as RulesetFunctionSchemaDefinition } from './rulesetFunction';
+export { assertValidRuleset, RulesetValidationError } from './validation/index';
+export { getDiagnosticSeverity } from './utils/severity';
+export { createRulesetFunction, SchemaDefinition as RulesetFunctionSchemaDefinition } from './function';
 export { Format } from './format';
 export { RulesetDefinition, RuleDefinition, ParserOptions, HumanReadableDiagnosticSeverity } from './types';
-export { Ruleset } from './ruleset';
-export { Rule } from './rule';
+export { Ruleset, StringifiedRuleset } from './ruleset';
+export { Rule, StringifiedRule } from './rule';